Move colonies to subfolder
