+++ shared/src/expansions/colonies/baseColonies.ts
@@ -1,8 +1,8 @@
+import { Resource, SymbolType } from '@shared/cards'
+import { PLAYER_RESOURCE_TO_PRODUCTION } from '@shared/constants'
+import { Colony, GridCellContent } from '@shared/game'
 import { discardCardsAction } from '@shared/player-actions'
-import { Resource, SymbolType } from '../cards'
-import { PLAYER_RESOURCE_TO_PRODUCTION } from '../constants'
-import { Colony, GridCellContent } from '../game'
-import { drawCard, drawCards, pushPendingAction, range } from '../utils'
+import { drawCard, drawCards, pushPendingAction, range } from '@shared/utils'
 import { colonyDrawCardsBonus } from './bonuses/colonyDrawCardsBonus'
 import { colonyResourceBonus } from './bonuses/colonyResourceBonus'
 import { colonyTileBonus } from './bonuses/colonyTileBonus'
